requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
PyYAML==6.0.2
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
sentence-transformers==3.2.1
pypdf==5.0.1
huggingface_hub==0.25.2

